extract benchmarkState to separate file
