+++ flutter/lib/ui/result_screen.dart
@@ -7,6 +7,7 @@ import 'package:share_plus/share_plus.dart';
 
 import 'package:mlperfbench/app_constants.dart';
 import 'package:mlperfbench/benchmark/benchmark.dart';
+import 'package:mlperfbench/benchmark/state.dart';
 import 'package:mlperfbench/icons.dart' as app_icons;
 import 'package:mlperfbench/info.dart';
 import 'package:mlperfbench/localizations/app_localizations.dart';
